Reimplement Quicksort Method in main.rs
To get to the bottom of the issues Isabelle and I were having with quicksort, the method and its helpers were rebuilt from the ground up. To better capture what was changed sequentially (instead of submitting one big edit), the next couple of commits will consists of the methods added/modified.

Co-authored-by: @IsabelleHjelden <hjeld013@morris.umn.edu>