Add mini_quicksort Helper Function to main.rs
In this commit, a helper function called mini_quicksort is added to main.rs to contain the call to partitioning the arrays to be sorted by two subsequent, recursive calls to mini_quicksort. The main reason why the logic of this method isn't included in the main quicksort method is to help with debugging the issues we were having in the lab.

Co-authored-by: @IsabelleHjelden <hjeld013@morris.umn.edu>