Add partition Helper Function for Quicksort in main.rs
To wrap up our work on quicksort, a new partitioning function utilizing a generic loop with nested while loops was added. The loops are used to increment values of i (a lower index) and j (a higher index) independently from one another as swapping is done to sort the method. At the very end, the value for i is returned to give mini_quicksort its new pivot for each call.

Co-authored-by: @IsabelleHjelden <hjeld013@morris.umn.edu>